+++ ifs-web-service/ifs-competition-mgt-service/src/test/java/org/innovateuk/ifs/management/controller/CompetitionManagementPanelInviteAssessorsControllerTest.java
@@ -421,7 +421,7 @@ public class CompetitionManagementPanelInviteAssessorsControllerTest extends Bas
             assertEquals(inviteOverviewResource.getBusinessType(), overviewAssessorRowViewModel.getBusinessType());
             assertEquals(inviteOverviewResource.getStatus(), overviewAssessorRowViewModel.getStatus());
             assertEquals(inviteOverviewResource.getDetails(), overviewAssessorRowViewModel.getDetails());
-            assertEquals((long) inviteOverviewResource.getInviteId(), overviewAssessorRowViewModel.getInviteId());
+            assertEquals(inviteOverviewResource.getInviteId(), overviewAssessorRowViewModel.getInviteId());
         });
     }
